# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": { "title": "CashControl API", "version": "1.0.0", "description": "Oracle 19c, JWT, roles y gastos idempotentes." },
  "servers": [{ "url": "/" }],
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/api/health": {
      "get": { "summary": "Salud pública de API y Oracle", "security": [],
        "responses": { "200": { "description": "Oracle CONNECTED" }, "500": { "$ref": "#/components/responses/ServerError" } } }
    },
    "/api/auth/register": {
      "post": { "summary": "Registrar usuario con rol user", "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Register" } } } },
        "responses": { "201": { "description": "Usuario registrado, sin tokens", "content": { "application/json": { "schema": {
          "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "$ref": "#/components/schemas/User" } }
        } } } }, "409": { "$ref": "#/components/responses/Conflict" }, "422": { "$ref": "#/components/responses/Validation" },
          "429": { "$ref": "#/components/responses/RateLimit" }, "500": { "$ref": "#/components/responses/ServerError" } }
      }
    },
    "/api/auth/login": {
      "post": { "summary": "Login con contraseña verificada por scrypt", "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Login" } } } },
        "responses": { "200": { "$ref": "#/components/responses/Session" }, "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/Validation" }, "429": { "$ref": "#/components/responses/RateLimit" },
          "500": { "$ref": "#/components/responses/ServerError" } }
      }
    },
    "/api/auth/refresh": {
      "post": { "summary": "Rotar refresh token de un solo uso", "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["refreshToken"], "properties": { "refreshToken": { "type": "string", "maxLength": 4096, "writeOnly": true } }
        } } } },
        "responses": { "200": { "$ref": "#/components/responses/Session" }, "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimit" }, "500": { "$ref": "#/components/responses/ServerError" } }
      }
    },
    "/api/auth/me": {
      "get": { "summary": "Perfil de la sesión autenticada", "responses": { "200": { "description": "Perfil público sin hash de contraseña" },
        "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" },
        "500": { "$ref": "#/components/responses/ServerError" } } }
    },
    "/api/admin/users": {
      "get": { "summary": "Listado administrativo de usuarios (solo admin, hasta 100)",
        "responses": { "200": { "description": "IDs, nombres y roles; sin hashes, contraseñas ni tokens" },
          "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" } }
      }
    },
    "/api/gastos": {
      "get": { "summary": "Listar gastos del usuario del JWT", "description": "Sin page/limit conserva el listado completo para sincronización Flutter.",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } }
        ],
        "responses": { "200": { "description": "Gastos y metadatos pagination si se solicitan", "content": { "application/json": { "schema": {
          "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/Expense" } },
            "pagination": { "type": "object", "properties": { "page": { "type": "integer" }, "limit": { "type": "integer" }, "total": { "type": "integer" }, "pages": { "type": "integer" } } }
          }
        } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" },
          "422": { "$ref": "#/components/responses/Validation" }, "500": { "$ref": "#/components/responses/ServerError" } }
      },
      "post": { "summary": "Crear/sincronizar gasto idempotente propio", "requestBody": { "required": true, "content": { "application/json": { "schema": {
        "allOf": [{ "$ref": "#/components/schemas/ExpenseInput" }, { "type": "object", "required": ["client_operation_id", "user_id", "updated_at"],
          "properties": { "client_operation_id": { "type": "string", "minLength": 1, "maxLength": 100 },
            "user_id": { "type": "string", "description": "Debe coincidir con userId del login", "maxLength": 100 },
            "updated_at": { "type": "string", "format": "date-time", "description": "ISO UTC con milisegundos" } }
        }]
      } } } }, "responses": { "201": { "$ref": "#/components/responses/Expense" }, "200": { "$ref": "#/components/responses/Expense" },
        "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" },
        "409": { "$ref": "#/components/responses/Conflict" }, "422": { "$ref": "#/components/responses/Validation" },
        "500": { "$ref": "#/components/responses/ServerError" } } }
    },
    "/api/gastos/{expenseId}": {
      "parameters": [{ "name": "expenseId", "in": "path", "required": true, "schema": { "type": "integer", "minimum": 1 } }],
      "get": { "summary": "Consultar gasto propio por server_id", "responses": { "200": { "$ref": "#/components/responses/Expense" },
        "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" },
        "404": { "$ref": "#/components/responses/NotFound" }, "422": { "$ref": "#/components/responses/Validation" }, "500": { "$ref": "#/components/responses/ServerError" } } },
      "put": { "summary": "Reemplazar campos editables de un gasto propio", "requestBody": { "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExpenseInput" } } } },
        "responses": { "200": { "$ref": "#/components/responses/Expense" }, "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/Validation" }, "500": { "$ref": "#/components/responses/ServerError" } }
      },
      "delete": { "summary": "Eliminar físicamente gasto propio (no purga SQLite)", "responses": { "200": { "description": "success, data.server_id y message" },
        "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" },
        "404": { "$ref": "#/components/responses/NotFound" }, "422": { "$ref": "#/components/responses/Validation" }, "500": { "$ref": "#/components/responses/ServerError" } } }
    }
  },
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } },
    "schemas": {
      "Login": { "type": "object", "required": ["username", "password"], "properties": {
        "username": { "type": "string", "minLength": 3, "maxLength": 100, "pattern": "^[a-zA-Z0-9._-]+$" },
        "password": { "type": "string", "minLength": 1, "maxLength": 128, "writeOnly": true }
      } },
      "Register": { "type": "object", "required": ["username", "email", "password"], "properties": {
        "username": { "type": "string", "minLength": 3, "maxLength": 100, "pattern": "^[a-zA-Z0-9._-]+$" },
        "email": { "type": "string", "format": "email", "maxLength": 254 },
        "password": { "type": "string", "minLength": 12, "maxLength": 128, "writeOnly": true },
        "role": { "type": "string", "enum": ["user"], "default": "user" }
      } },
      "User": { "type": "object", "properties": { "id": { "type": "string" }, "username": { "type": "string" },
        "email": { "type": "string", "nullable": true }, "role": { "type": "string", "enum": ["user", "admin"] } } },
      "Session": { "type": "object", "required": ["success", "userId", "accessToken", "refreshToken"], "properties": {
        "success": { "type": "boolean" }, "userId": { "type": "string" }, "user": { "$ref": "#/components/schemas/User" },
        "accessToken": { "type": "string" }, "refreshToken": { "type": "string" }
      } },
      "ExpenseInput": { "type": "object", "required": ["category_id", "amount", "description", "date"], "properties": {
        "category_id": { "type": "string", "minLength": 1, "maxLength": 100 },
        "amount": { "type": "number", "exclusiveMinimum": true, "minimum": 0, "maximum": 9999999999.99, "multipleOf": 0.01 },
        "description": { "type": "string", "minLength": 1, "maxLength": 255 },
        "date": { "type": "string", "format": "date-time", "description": "ISO UTC con milisegundos" },
        "latitude": { "type": "number", "minimum": -90, "maximum": 90, "nullable": true },
        "longitude": { "type": "number", "minimum": -180, "maximum": 180, "nullable": true }
      } },
      "Expense": { "allOf": [{ "$ref": "#/components/schemas/ExpenseInput" }, { "type": "object", "properties": {
        "server_id": { "type": "integer" }, "client_operation_id": { "type": "string" }, "user_id": { "type": "string" },
        "created_at": { "type": "string", "format": "date-time" }, "updated_at": { "type": "string", "format": "date-time" }
      } }] },
      "Error": { "type": "object", "required": ["success", "message"], "properties": {
        "success": { "type": "boolean", "enum": [false] }, "message": { "type": "string" },
        "errors": { "type": "object", "additionalProperties": { "type": "string" } }
      } }
    },
    "responses": {
      "Session": { "description": "Sesión JWT; no registrar tokens en logs", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Session" } } } },
      "Expense": { "description": "success y data con el gasto", "content": { "application/json": { "schema": { "type": "object", "properties": {
        "success": { "type": "boolean" }, "data": { "$ref": "#/components/schemas/Expense" }
      } } } } },
      "Unauthorized": { "description": "401: token ausente, inválido, vencido, usado o credenciales incorrectas", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Forbidden": { "description": "403: rol autenticado sin permiso", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "404: recurso inexistente o ajeno (no revelar su existencia)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Conflict": { "description": "409: conflicto de unicidad", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Validation": { "description": "422: errores asociados a campos", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "RateLimit": { "description": "429: demasiados intentos, Retry-After en segundos", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "ServerError": { "description": "500: error interno sin trazas", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    }
  }
}