# backend/server.js
require('dotenv').config({ path: require('node:path').join(__dirname, '.env'), quiet: true });

const express = require('express');
const oracledb = require('oracledb');
const https = require('node:https');
const fs = require('node:fs');
const { AuthStore } = require('./auth_store');
const { createAuth } = require('./auth_routes');
const { validateExpense, parsePagination, validExpenseId } = require('./expense_validation');

const app = express();
const PORT = process.env.PORT || 3000;
const JWT_SECRET = process.env.JWT_SECRET;
const ACCESS_TOKEN_TTL = process.env.ACCESS_TOKEN_TTL || '15m';
const REFRESH_TOKEN_TTL = process.env.REFRESH_TOKEN_TTL || '7d';

if (!JWT_SECRET || Buffer.byteLength(JWT_SECRET) < 32) {
  throw new Error('JWT_SECRET debe configurarse con al menos 32 bytes en backend/.env');
}

app.disable('x-powered-by');
app.use(express.json({ limit: '32kb' }));
app.use((req, res, next) => {
  res.set('Cache-Control', 'no-store');
  res.set('X-Content-Type-Options', 'nosniff');
  next();
});
app.get('/api/openapi.json', (req, res) => res.json(require('./openapi.json')));
const auth = createAuth({ store: new AuthStore(), secret: JWT_SECRET,
  accessTtl: ACCESS_TOKEN_TTL, refreshTtl: REFRESH_TOKEN_TTL });
const requireAuth = auth.requireAuth;
app.use('/api', auth.router);
app.use('/api/gastos', requireAuth, auth.requireRole('user', 'admin'));

async function ensureExpensesTable(connection) {
  try {
    await connection.execute(`
      CREATE TABLE CC_GASTOS_SYNC (
        ID_GASTO NUMBER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
        CLIENT_OPERATION_ID VARCHAR2(100) NOT NULL UNIQUE,
        USER_ID VARCHAR2(100) NOT NULL,
        CATEGORY_ID VARCHAR2(100) NOT NULL,
        AMOUNT NUMBER(12,2) NOT NULL,
        DESCRIPTION VARCHAR2(255) NOT NULL,
        EXPENSE_DATE TIMESTAMP NOT NULL,
        CREATED_AT TIMESTAMP DEFAULT SYSTIMESTAMP NOT NULL,
        UPDATED_AT TIMESTAMP DEFAULT SYSTIMESTAMP NOT NULL,
        LATITUDE NUMBER(10,6),
        LONGITUDE NUMBER(10,6)
      )
    `);
  } catch (error) {
    if (error.errorNum !== 955) throw error;
    // Table already existed from a previous week: add the Week 14 location columns if missing.
    for (const column of ['LATITUDE NUMBER(10,6)', 'LONGITUDE NUMBER(10,6)']) {
      try {
        await connection.execute(`ALTER TABLE CC_GASTOS_SYNC ADD (${column})`);
      } catch (alterError) {
        if (alterError.errorNum !== 1430) throw alterError;
      }
    }
  }
}

function expenseFromRow(row) {
  return {
    server_id: row.ID_GASTO,
    client_operation_id: row.CLIENT_OPERATION_ID,
    user_id: row.USER_ID,
    category_id: row.CATEGORY_ID,
    amount: row.AMOUNT,
    description: row.DESCRIPTION,
    date: row.EXPENSE_DATE.toISOString(),
    created_at: row.CREATED_AT.toISOString(),
    updated_at: row.UPDATED_AT.toISOString(),
    latitude: row.LATITUDE ?? null,
    longitude: row.LONGITUDE ?? null
  };
}

app.get('/api/gastos', requireAuth, async (req, res) => {
  const pagination = parsePagination(req.query);
  if (pagination === false) {
    return res.status(422).json({ success: false, message: 'Error de validación', errors: {
      pagination: 'page debe ser positivo y limit debe estar entre 1 y 100'
    } });
  }
  let connection;
  try {
    connection = await oracledb.getConnection({
      user: process.env.DB_USER,
      password: process.env.DB_PASSWORD,
      connectString: process.env.DB_CONNECT_STRING
    });
    await ensureExpensesTable(connection);
    const result = await connection.execute(
      `SELECT * FROM CC_GASTOS_SYNC WHERE USER_ID = :userId ORDER BY EXPENSE_DATE DESC, ID_GASTO DESC` +
        (pagination ? ' OFFSET :offset ROWS FETCH NEXT :limit ROWS ONLY' : ''),
      { userId: req.userId, ...(pagination ? { offset: pagination.offset, limit: pagination.limit } : {}) },
      { outFormat: oracledb.OUT_FORMAT_OBJECT }
    );
    let metadata;
    if (pagination) {
      const count = await connection.execute(
        'SELECT COUNT(*) AS TOTAL FROM CC_GASTOS_SYNC WHERE USER_ID = :userId',
        { userId: req.userId }, { outFormat: oracledb.OUT_FORMAT_OBJECT }
      );
      const total = count.rows[0].TOTAL;
      metadata = { page: pagination.page, limit: pagination.limit, total, pages: Math.ceil(total / pagination.limit) };
    }
    res.json({ success: true, data: result.rows.map(expenseFromRow), ...(metadata ? { pagination: metadata } : {}) });
  } catch (error) {
    console.error('Error al listar gastos:', error.message);
    res.status(500).json({ success: false, message: 'No fue posible listar los gastos' });
  } finally {
    if (connection) await connection.close();
  }
});

app.post('/api/gastos', requireAuth, async (req, res) => {
  let connection;
  const { client_operation_id: operationId, user_id: userId, category_id: categoryId,
    amount, description, date, updated_at: updatedAt, latitude, longitude } = req.body || {};
  const errors = validateExpense(req.body, { userId: req.userId });
  if (Object.keys(errors).length) {
    return res.status(422).json({ success: false, message: 'Error de validación', errors });
  }
  try {
    connection = await oracledb.getConnection({
      user: process.env.DB_USER,
      password: process.env.DB_PASSWORD,
      connectString: process.env.DB_CONNECT_STRING
    });
    await ensureExpensesTable(connection);
    const existing = await connection.execute(
      'SELECT USER_ID FROM CC_GASTOS_SYNC WHERE CLIENT_OPERATION_ID = :operationId',
      { operationId }, { outFormat: oracledb.OUT_FORMAT_OBJECT }
    );
    if (existing.rows.length && existing.rows[0].USER_ID !== req.userId) {
      return res.status(409).json({ success: false, message: 'Identificador de operación no disponible' });
    }
    await connection.execute(`
      MERGE INTO CC_GASTOS_SYNC target
      USING (SELECT :operationId client_operation_id FROM dual) source
      ON (target.CLIENT_OPERATION_ID = source.client_operation_id AND target.USER_ID = :userId)
      WHEN MATCHED THEN UPDATE SET
        target.CATEGORY_ID = :categoryId,
        target.AMOUNT = :amount,
        target.DESCRIPTION = :description,
        target.EXPENSE_DATE = :expenseDate,
        target.UPDATED_AT = :updatedAt,
        target.LATITUDE = :latitude,
        target.LONGITUDE = :longitude
        WHERE target.UPDATED_AT < :updatedAt
      WHEN NOT MATCHED THEN INSERT
        (CLIENT_OPERATION_ID, USER_ID, CATEGORY_ID, AMOUNT, DESCRIPTION, EXPENSE_DATE, UPDATED_AT, LATITUDE, LONGITUDE)
        VALUES (:operationId, :userId, :categoryId, :amount, :description,
          :expenseDate, :updatedAt, :latitude, :longitude)
    `, { operationId, userId: String(userId), categoryId: String(categoryId), amount,
      description: description || '', expenseDate: new Date(date), updatedAt: new Date(updatedAt),
      latitude: latitude ?? null, longitude: longitude ?? null }, { autoCommit: true });
    const result = await connection.execute(
      `SELECT * FROM CC_GASTOS_SYNC WHERE CLIENT_OPERATION_ID = :operationId AND USER_ID = :userId`,
      { operationId, userId: req.userId }, { outFormat: oracledb.OUT_FORMAT_OBJECT }
    );
    res.status(existing.rows.length ? 200 : 201).json({ success: true, data: expenseFromRow(result.rows[0]) });
  } catch (error) {
    if (error.errorNum === 1) {
      return res.status(409).json({ success: false, message: 'Identificador de operación duplicado; consulte el registro antes de reintentar' });
    }
    console.error('Error al sincronizar gasto:', error.message);
    res.status(500).json({ success: false, message: 'No fue posible sincronizar el gasto' });
  } finally {
    if (connection) await connection.close();
  }
});

app.param('expenseId', (req, res, next, value) => {
  if (!validExpenseId(value)) {
    return res.status(422).json({ success: false, message: 'Error de validación', errors: { expenseId: 'ID inválido' } });
  }
  next();
});

app.get('/api/gastos/:expenseId', requireAuth, async (req, res) => {
  let connection;
  try {
    connection = await oracledb.getConnection({ user: process.env.DB_USER, password: process.env.DB_PASSWORD, connectString: process.env.DB_CONNECT_STRING });
    await ensureExpensesTable(connection);
    const result = await connection.execute(
      'SELECT * FROM CC_GASTOS_SYNC WHERE ID_GASTO = :expenseId AND USER_ID = :userId',
      { expenseId: Number(req.params.expenseId), userId: req.userId }, { outFormat: oracledb.OUT_FORMAT_OBJECT }
    );
    if (!result.rows.length) return res.status(404).json({ success: false, message: 'Gasto no encontrado' });
    res.json({ success: true, data: expenseFromRow(result.rows[0]) });
  } catch (_) {
    res.status(500).json({ success: false, message: 'No fue posible consultar el gasto' });
  } finally {
    if (connection) await connection.close();
  }
});

app.put('/api/gastos/:expenseId', requireAuth, async (req, res) => {
  const errors = validateExpense(req.body, { create: false, userId: req.userId });
  if (Object.keys(errors).length) return res.status(422).json({ success: false, message: 'Error de validación', errors });
  let connection;
  try {
    connection = await oracledb.getConnection({ user: process.env.DB_USER, password: process.env.DB_PASSWORD, connectString: process.env.DB_CONNECT_STRING });
    await ensureExpensesTable(connection);
    const result = await connection.execute(
      `UPDATE CC_GASTOS_SYNC SET CATEGORY_ID = :categoryId, AMOUNT = :amount, DESCRIPTION = :description,
        EXPENSE_DATE = :expenseDate, LATITUDE = :latitude, LONGITUDE = :longitude, UPDATED_AT = SYSTIMESTAMP
        WHERE ID_GASTO = :expenseId AND USER_ID = :userId`,
      { categoryId: req.body.category_id, amount: req.body.amount, description: req.body.description,
        expenseDate: new Date(req.body.date), latitude: req.body.latitude ?? null, longitude: req.body.longitude ?? null,
        expenseId: Number(req.params.expenseId), userId: req.userId }, { autoCommit: false }
    );
    if (!result.rowsAffected) return res.status(404).json({ success: false, message: 'Gasto no encontrado' });
    const updated = await connection.execute(
      'SELECT * FROM CC_GASTOS_SYNC WHERE ID_GASTO = :expenseId AND USER_ID = :userId',
      { expenseId: Number(req.params.expenseId), userId: req.userId }, { outFormat: oracledb.OUT_FORMAT_OBJECT }
    );
    await connection.commit();
    res.json({ success: true, data: expenseFromRow(updated.rows[0]) });
  } catch (_) {
    res.status(500).json({ success: false, message: 'No fue posible actualizar el gasto' });
  } finally {
    if (connection) await connection.close();
  }
});

app.delete('/api/gastos/:expenseId', requireAuth, async (req, res) => {
  let connection;
  try {
    connection = await oracledb.getConnection({ user: process.env.DB_USER, password: process.env.DB_PASSWORD, connectString: process.env.DB_CONNECT_STRING });
    await ensureExpensesTable(connection);
    const result = await connection.execute(
      'DELETE FROM CC_GASTOS_SYNC WHERE ID_GASTO = :expenseId AND USER_ID = :userId',
      { expenseId: Number(req.params.expenseId), userId: req.userId }, { autoCommit: true }
    );
    if (!result.rowsAffected) return res.status(404).json({ success: false, message: 'Gasto no encontrado' });
    res.json({ success: true, data: { server_id: Number(req.params.expenseId) }, message: 'Gasto eliminado' });
  } catch (_) {
    res.status(500).json({ success: false, message: 'No fue posible eliminar el gasto' });
  } finally {
    if (connection) await connection.close();
  }
});

app.get('/api/health', async (req, res) => {
  let connection;

  try {
    connection = await oracledb.getConnection({
      user: process.env.DB_USER,
      password: process.env.DB_PASSWORD,
      connectString: process.env.DB_CONNECT_STRING
    });

    const result = await connection.execute(
      `SELECT 'CONNECTED' AS STATUS FROM DUAL`
    );

    res.status(200).json({
      success: true,
      message: 'API de CashControl funcionando correctamente',
      database: result.rows[0][0]
    });
  } catch (error) {
    console.error('Error de conexión con Oracle:', error.message);

    res.status(500).json({
      success: false,
      message: 'La API está funcionando, pero no fue posible conectar con Oracle'
    });
  } finally {
    if (connection) {
      try {
        await connection.close();
      } catch (error) {
        console.error('Error al cerrar la conexión:', error.message);
      }
    }
  }
});

app.use((error, req, res, next) => {
  if (res.headersSent) return next(error);
  const malformed = error instanceof SyntaxError && error.status === 400;
  res.status(malformed ? 400 : 500).json({ success: false,
    message: malformed ? 'JSON inválido' : 'Ocurrió un error en el servidor' });
});

if (require.main === module) {
  const certificate = process.env.HTTPS_CERT_FILE;
  const key = process.env.HTTPS_KEY_FILE;
  if (process.env.NODE_ENV === 'production' && (!certificate || !key)) {
    throw new Error('Producción requiere HTTPS_CERT_FILE y HTTPS_KEY_FILE');
  }
  if (!!certificate !== !!key) throw new Error('Configure ambas rutas de certificado y clave TLS');
  const server = certificate ? https.createServer({ cert: fs.readFileSync(certificate), key: fs.readFileSync(key) }, app) : app;
  server.listen(PORT, () => {
    console.log(`Servidor CashControl ejecutándose en ${certificate ? 'https' : 'http'}://localhost:${PORT}`);
  });
}

module.exports = app;